Refactor form group tests to enhance nullable control handling and apply control option properties
